lisa.analysis: Keep all rt-app related analysis in the same module
Splitting across module is not really needed, since they achieve similar things.
