+++ tests/test_wlgen_rtapp.py
@@ -20,7 +20,7 @@ import json
 import os
 
 from lisa.wlgen.rta import RTA, Periodic, Ramp, Step, RunAndSync
-from lisa.analysis.rta_logs import PerfAnalysis
+from lisa.analysis.rta import PerfAnalysis
 
 from .utils import StorageTestCase, create_local_target, ASSET_DIR
